requirements.txt
requests
python-dotenv
huesdk
python-firebase
BlynkLib
sense-hat
firebase-admin